refactor: move getReferences to utils
