+++ core/src/index.ts
@@ -2,3 +2,4 @@ export * as models from "./models";
 export * as types from "./types";
 export * as config from "./config";
 export * as repositories from "./repositories";
+export * as utils from "./utils";
