+++ core/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./makeGPromise";
+export * from "./getReferences";
